Extracted superclass to reduce code repetition
