+++ lib/api/v3/database_checks/has_many_association_any.rb
@@ -1,17 +1,11 @@
 module Api
   module V3
     module DatabaseChecks
-      class HasManyAssociationAny
+      class HasManyAssociationAny < GenericCheck
         def initialize(object, options)
-          @object = object
+          super(object, options)
           @association = options[:association]
           @conditions = options[:conditions]
-          @link = options[:link]
-        end
-
-        def call(report_status)
-          return if passing?
-          report_status.add_error(error)
         end
 
         def passing?
@@ -25,13 +19,10 @@ module Api
         private
 
         def error
-          {
-            object: @object,
-            type: self.class.name,
+          super.merge({
             message: "#{@association} missing #{@conditions}",
-            suggestion: "Please create some at :link",
-            link: @link
-          }
+            suggestion: "Please create some at :link"
+          })
         end
       end
     end
